requirements-macbook-without-nvdia-gpu.txt
contourpy==1.3.1
cramjam==2.9.1
cycler==0.12.1
duckdb==0.9.2
fastparquet==2024.2.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pandas==2.2.1
pathlib==1.0.1
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scikit-learn==1.5.0
scipy==1.12.0
seaborn==0.13.2
six==1.17.0
sympy==1.13.3
threadpoolctl==3.5.0
toml==0.10.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
